f8-FE-Chat - refactor Chat, tach thanh 2 file  RestAPI, Websocket
